Move base project data out of seed utils and into generic data
